Create an Animated Plot of Survival Curves of CIFs as a Function of a Continuous Variable

Using a previously fit time-to-event model, this function plots a single covariate-specific survival curve or CIF. Using either the plotly or the gganimate package, an animated plot is created that is either simply showing how the curve evolves over values of the continuous covariate, or includes a slide to let the user interact with it.

Usage

plot_surv_animated(time, status, variable, group=NULL,
                   data, model, cif=FALSE,
                   na.action=options()$na.action,
                   horizon=NULL, fixed_t=NULL, max_t=Inf,
                   slider=TRUE, size=1, color="black",
                   linetype="solid", alpha=1,
                   xlab="Time", ylab="Survival Probability",
                   title=NULL, subtitle=NULL,
                   gg_theme=ggplot2::theme_bw(),
                   facet_args=list(), ...)

Arguments

time

A single character string specifying the time-to-event variable. Needs to be a valid column name of a numeric variable in data.

status

A single character string specifying the status variable, indicating if a person has experienced an event or not. Needs to be a valid column name of a numeric or logical variable in data.

variable

A single character string specifying the continuous variable of interest, for which the survival curves should be estimated. This variable has to be contained in the data.frame that is supplied to the data argument.

group

An optional single character string specifying a factor variable in data. When used, the plot is created conditional on this factor variable, meaning that a facetted plot is produced with one facet for each level of the factor variable. See curve_cont for a detailed description of the estimation strategy. Set to NULL (default) to use no grouping variable.

data

A data.frame containing all required variables.

model

A model describing the time-to-event process (such as an coxph model). Needs to include variable as an independent variable. It also has to have an associated predictRisk method. See ?predictRisk for more details.

cif

Whether to plot the cumulative incidence (CIF) instead of the survival probability. If multiple failure types are present, the survival probability cannot be estimated in an unbiased way. This function will always return CIF estimates in that case.

na.action

How missing values should be handled. Can be one of: na.fail, na.omit, na.pass, na.exclude or a user-defined custom function. Also accepts strings of the function names. See ?na.action for more details. By default it uses the na.action which is set in the global options by the respective user.

horizon

A numeric vector containing a range of values of variable for which the survival curves should be calculated or NULL (default). If NULL, the horizon is constructed as a sequence from the lowest to the highest value observed in variable with 40 equally spaced steps.

fixed_t

A numeric vector containing points in time at which the survival probabilities should be calculated or NULL (default). If NULL, the survival probability is estimated at every point in time at which an event occured.

max_t

A number indicating the latest survival time which is to be plotted.

slider

Whether to include a slider that controls the value of the continuous covariate or not. If not, the plot simply cycles through the values in horizon at a constant speed.

size

A single number specifying the size of the drawn curve.

color

A single character string specifying the color of the curve.

linetype

A single character string specifying the linetype of the curve.

alpha

The transparency level of the plot.

xlab

A character string used as the x-axis label of the plot.

ylab

A character string used as the y-axis label of the plot.

title

A character string used as the title of the plot. When slider=FALSE is used, the title is also used in a dynamic way to show which value the continuous variable currently has. The user-defined title is still shown, but the covariate level is also appended to it in that case.

subtitle

A character string used as the subtitle of the plot.

gg_theme

A ggplot2 theme which is applied to the plot.

facet_args

A named list of arguments that are passed to the facet_wrap function call when creating a plot separated by groups. Ignored if group=NULL. Any argument except the facets argument of the facet_wrap function can be used. For example, if the user wants to allow free y-scales, this argument could be set to list(scales="free_y").

...

Further arguments passed to curve_cont.

Details

When the plot doesn't have to be on paper, it is also possible to use animation to represent how the survival curve or CIF changes when different values of a continuous covariate are used. There are two options included in this function. The first one is creating a plot with a slider that can be interactively manipulated by the user (argument slider=TRUE). The second option creates a .gif file in which the continuous variable changes constantly over time (argument slider=FALSE).

This type of plot is only usefull when visualizing the effect on a computer. It is also important to keep in mind that these plots might look slightly different to users with different hardware and or software.

An additional interactive plot can be created using the plot_surv_3Dsurface function with interactive=TRUE.

Value

Returns a plotly object if slider=TRUE and a gganim object if slider=FALSE.

Author

Robin Denz

Examples

library(contsurvplot)
library(riskRegression)
library(survival)
library(splines)
library(ggplot2)
library(gganimate)
#> No renderer backend detected. gganimate will default to writing frames to separate files
#> Consider installing:
#> - the `gifski` package for gif output
#> - the `av` package for video output
#> and restarting the R session
library(transformr)
library(plotly)

# using data from the survival package
data(nafld, package="survival")

# take a random sample to keep example fast
set.seed(42)
nafld1 <- nafld1[sample(nrow(nafld1), 150), ]

# fit cox-model with age
model <- coxph(Surv(futime, status) ~ age, data=nafld1, x=TRUE)

if (interactive()) {
# plot effect of age on survival for ages 50 to 80
plot_surv_animated(time="futime",
                   status="status",
                   variable="age",
                   data=nafld1,
                   model=model,
                   horizon=seq(50, 80, 1),
                   slider=TRUE)

# plot effect of age on survival using an interactive plot for ages 50 to 80
plot_surv_animated(time="futime",
                   status="status",
                   variable="age",
                   data=nafld1,
                   model=model,
                   horizon=seq(50, 80, 1),
                   slider=FALSE)
}
#>   [1] "./gganim_plot0001.png" "./gganim_plot0002.png" "./gganim_plot0003.png"
#>   [4] "./gganim_plot0004.png" "./gganim_plot0005.png" "./gganim_plot0006.png"
#>   [7] "./gganim_plot0007.png" "./gganim_plot0008.png" "./gganim_plot0009.png"
#>  [10] "./gganim_plot0010.png" "./gganim_plot0011.png" "./gganim_plot0012.png"
#>  [13] "./gganim_plot0013.png" "./gganim_plot0014.png" "./gganim_plot0015.png"
#>  [16] "./gganim_plot0016.png" "./gganim_plot0017.png" "./gganim_plot0018.png"
#>  [19] "./gganim_plot0019.png" "./gganim_plot0020.png" "./gganim_plot0021.png"
#>  [22] "./gganim_plot0022.png" "./gganim_plot0023.png" "./gganim_plot0024.png"
#>  [25] "./gganim_plot0025.png" "./gganim_plot0026.png" "./gganim_plot0027.png"
#>  [28] "./gganim_plot0028.png" "./gganim_plot0029.png" "./gganim_plot0030.png"
#>  [31] "./gganim_plot0031.png" "./gganim_plot0032.png" "./gganim_plot0033.png"
#>  [34] "./gganim_plot0034.png" "./gganim_plot0035.png" "./gganim_plot0036.png"
#>  [37] "./gganim_plot0037.png" "./gganim_plot0038.png" "./gganim_plot0039.png"
#>  [40] "./gganim_plot0040.png" "./gganim_plot0041.png" "./gganim_plot0042.png"
#>  [43] "./gganim_plot0043.png" "./gganim_plot0044.png" "./gganim_plot0045.png"
#>  [46] "./gganim_plot0046.png" "./gganim_plot0047.png" "./gganim_plot0048.png"
#>  [49] "./gganim_plot0049.png" "./gganim_plot0050.png" "./gganim_plot0051.png"
#>  [52] "./gganim_plot0052.png" "./gganim_plot0053.png" "./gganim_plot0054.png"
#>  [55] "./gganim_plot0055.png" "./gganim_plot0056.png" "./gganim_plot0057.png"
#>  [58] "./gganim_plot0058.png" "./gganim_plot0059.png" "./gganim_plot0060.png"
#>  [61] "./gganim_plot0061.png" "./gganim_plot0062.png" "./gganim_plot0063.png"
#>  [64] "./gganim_plot0064.png" "./gganim_plot0065.png" "./gganim_plot0066.png"
#>  [67] "./gganim_plot0067.png" "./gganim_plot0068.png" "./gganim_plot0069.png"
#>  [70] "./gganim_plot0070.png" "./gganim_plot0071.png" "./gganim_plot0072.png"
#>  [73] "./gganim_plot0073.png" "./gganim_plot0074.png" "./gganim_plot0075.png"
#>  [76] "./gganim_plot0076.png" "./gganim_plot0077.png" "./gganim_plot0078.png"
#>  [79] "./gganim_plot0079.png" "./gganim_plot0080.png" "./gganim_plot0081.png"
#>  [82] "./gganim_plot0082.png" "./gganim_plot0083.png" "./gganim_plot0084.png"
#>  [85] "./gganim_plot0085.png" "./gganim_plot0086.png" "./gganim_plot0087.png"
#>  [88] "./gganim_plot0088.png" "./gganim_plot0089.png" "./gganim_plot0090.png"
#>  [91] "./gganim_plot0091.png" "./gganim_plot0092.png" "./gganim_plot0093.png"
#>  [94] "./gganim_plot0094.png" "./gganim_plot0095.png" "./gganim_plot0096.png"
#>  [97] "./gganim_plot0097.png" "./gganim_plot0098.png" "./gganim_plot0099.png"
#> [100] "./gganim_plot0100.png"
#> attr(,"frame_vars")
#>     frame nframes progress frame_time
#> 1       1     100     0.01   50.00000
#> 2       2     100     0.02   50.30303
#> 3       3     100     0.03   50.60606
#> 4       4     100     0.04   50.90909
#> 5       5     100     0.05   51.21212
#> 6       6     100     0.06   51.51515
#> 7       7     100     0.07   51.81818
#> 8       8     100     0.08   52.12121
#> 9       9     100     0.09   52.42424
#> 10     10     100     0.10   52.72727
#> 11     11     100     0.11   53.03030
#> 12     12     100     0.12   53.33333
#> 13     13     100     0.13   53.63636
#> 14     14     100     0.14   53.93939
#> 15     15     100     0.15   54.24242
#> 16     16     100     0.16   54.54545
#> 17     17     100     0.17   54.84848
#> 18     18     100     0.18   55.15152
#> 19     19     100     0.19   55.45455
#> 20     20     100     0.20   55.75758
#> 21     21     100     0.21   56.06061
#> 22     22     100     0.22   56.36364
#> 23     23     100     0.23   56.66667
#> 24     24     100     0.24   56.96970
#> 25     25     100     0.25   57.27273
#> 26     26     100     0.26   57.57576
#> 27     27     100     0.27   57.87879
#> 28     28     100     0.28   58.18182
#> 29     29     100     0.29   58.48485
#> 30     30     100     0.30   58.78788
#> 31     31     100     0.31   59.09091
#> 32     32     100     0.32   59.39394
#> 33     33     100     0.33   59.69697
#> 34     34     100     0.34   60.00000
#> 35     35     100     0.35   60.30303
#> 36     36     100     0.36   60.60606
#> 37     37     100     0.37   60.90909
#> 38     38     100     0.38   61.21212
#> 39     39     100     0.39   61.51515
#> 40     40     100     0.40   61.81818
#> 41     41     100     0.41   62.12121
#> 42     42     100     0.42   62.42424
#> 43     43     100     0.43   62.72727
#> 44     44     100     0.44   63.03030
#> 45     45     100     0.45   63.33333
#> 46     46     100     0.46   63.63636
#> 47     47     100     0.47   63.93939
#> 48     48     100     0.48   64.24242
#> 49     49     100     0.49   64.54545
#> 50     50     100     0.50   64.84848
#> 51     51     100     0.51   65.15152
#> 52     52     100     0.52   65.45455
#> 53     53     100     0.53   65.75758
#> 54     54     100     0.54   66.06061
#> 55     55     100     0.55   66.36364
#> 56     56     100     0.56   66.66667
#> 57     57     100     0.57   66.96970
#> 58     58     100     0.58   67.27273
#> 59     59     100     0.59   67.57576
#> 60     60     100     0.60   67.87879
#> 61     61     100     0.61   68.18182
#> 62     62     100     0.62   68.48485
#> 63     63     100     0.63   68.78788
#> 64     64     100     0.64   69.09091
#> 65     65     100     0.65   69.39394
#> 66     66     100     0.66   69.69697
#> 67     67     100     0.67   70.00000
#> 68     68     100     0.68   70.30303
#> 69     69     100     0.69   70.60606
#> 70     70     100     0.70   70.90909
#> 71     71     100     0.71   71.21212
#> 72     72     100     0.72   71.51515
#> 73     73     100     0.73   71.81818
#> 74     74     100     0.74   72.12121
#> 75     75     100     0.75   72.42424
#> 76     76     100     0.76   72.72727
#> 77     77     100     0.77   73.03030
#> 78     78     100     0.78   73.33333
#> 79     79     100     0.79   73.63636
#> 80     80     100     0.80   73.93939
#> 81     81     100     0.81   74.24242
#> 82     82     100     0.82   74.54545
#> 83     83     100     0.83   74.84848
#> 84     84     100     0.84   75.15152
#> 85     85     100     0.85   75.45455
#> 86     86     100     0.86   75.75758
#> 87     87     100     0.87   76.06061
#> 88     88     100     0.88   76.36364
#> 89     89     100     0.89   76.66667
#> 90     90     100     0.90   76.96970
#> 91     91     100     0.91   77.27273
#> 92     92     100     0.92   77.57576
#> 93     93     100     0.93   77.87879
#> 94     94     100     0.94   78.18182
#> 95     95     100     0.95   78.48485
#> 96     96     100     0.96   78.78788
#> 97     97     100     0.97   79.09091
#> 98     98     100     0.98   79.39394
#> 99     99     100     0.99   79.69697
#> 100   100     100     1.00   80.00000
#>                                                                               frame_source
#> 1   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0001.png
#> 2   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0002.png
#> 3   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0003.png
#> 4   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0004.png
#> 5   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0005.png
#> 6   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0006.png
#> 7   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0007.png
#> 8   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0008.png
#> 9   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0009.png
#> 10  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0010.png
#> 11  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0011.png
#> 12  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0012.png
#> 13  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0013.png
#> 14  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0014.png
#> 15  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0015.png
#> 16  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0016.png
#> 17  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0017.png
#> 18  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0018.png
#> 19  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0019.png
#> 20  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0020.png
#> 21  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0021.png
#> 22  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0022.png
#> 23  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0023.png
#> 24  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0024.png
#> 25  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0025.png
#> 26  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0026.png
#> 27  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0027.png
#> 28  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0028.png
#> 29  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0029.png
#> 30  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0030.png
#> 31  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0031.png
#> 32  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0032.png
#> 33  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0033.png
#> 34  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0034.png
#> 35  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0035.png
#> 36  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0036.png
#> 37  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0037.png
#> 38  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0038.png
#> 39  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0039.png
#> 40  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0040.png
#> 41  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0041.png
#> 42  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0042.png
#> 43  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0043.png
#> 44  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0044.png
#> 45  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0045.png
#> 46  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0046.png
#> 47  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0047.png
#> 48  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0048.png
#> 49  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0049.png
#> 50  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0050.png
#> 51  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0051.png
#> 52  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0052.png
#> 53  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0053.png
#> 54  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0054.png
#> 55  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0055.png
#> 56  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0056.png
#> 57  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0057.png
#> 58  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0058.png
#> 59  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0059.png
#> 60  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0060.png
#> 61  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0061.png
#> 62  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0062.png
#> 63  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0063.png
#> 64  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0064.png
#> 65  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0065.png
#> 66  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0066.png
#> 67  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0067.png
#> 68  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0068.png
#> 69  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0069.png
#> 70  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0070.png
#> 71  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0071.png
#> 72  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0072.png
#> 73  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0073.png
#> 74  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0074.png
#> 75  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0075.png
#> 76  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0076.png
#> 77  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0077.png
#> 78  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0078.png
#> 79  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0079.png
#> 80  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0080.png
#> 81  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0081.png
#> 82  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0082.png
#> 83  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0083.png
#> 84  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0084.png
#> 85  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0085.png
#> 86  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0086.png
#> 87  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0087.png
#> 88  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0088.png
#> 89  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0089.png
#> 90  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0090.png
#> 91  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0091.png
#> 92  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0092.png
#> 93  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0093.png
#> 94  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0094.png
#> 95  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0095.png
#> 96  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0096.png
#> 97  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0097.png
#> 98  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0098.png
#> 99  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0099.png
#> 100 C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\14289a13797/gganim_plot0100.png

## showing non-linear effects

# fit cox-model with bmi modeled using B-Splines,
# adjusting for age and sex
model2 <- coxph(Surv(futime, status) ~ age + male + bs(bmi, df=3),
                data=nafld1, x=TRUE)

if (interactive()) {
# plot effect of bmi on survival using slider
plot_surv_animated(time="futime",
                   status="status",
                   variable="bmi",
                   data=nafld1,
                   model=model2,
                   slider=TRUE)

# plot effect of bmi on survival using a .gif
plot_surv_animated(time="futime",
                   status="status",
                   variable="bmi",
                   data=nafld1,
                   model=model2,
                   slider=FALSE)
}
#> Warning: file_renderer failed to copy frames to the destination directory
#>   [1] "./gganim_plot0001.png" "./gganim_plot0002.png" "./gganim_plot0003.png"
#>   [4] "./gganim_plot0004.png" "./gganim_plot0005.png" "./gganim_plot0006.png"
#>   [7] "./gganim_plot0007.png" "./gganim_plot0008.png" "./gganim_plot0009.png"
#>  [10] "./gganim_plot0010.png" "./gganim_plot0011.png" "./gganim_plot0012.png"
#>  [13] "./gganim_plot0013.png" "./gganim_plot0014.png" "./gganim_plot0015.png"
#>  [16] "./gganim_plot0016.png" "./gganim_plot0017.png" "./gganim_plot0018.png"
#>  [19] "./gganim_plot0019.png" "./gganim_plot0020.png" "./gganim_plot0021.png"
#>  [22] "./gganim_plot0022.png" "./gganim_plot0023.png" "./gganim_plot0024.png"
#>  [25] "./gganim_plot0025.png" "./gganim_plot0026.png" "./gganim_plot0027.png"
#>  [28] "./gganim_plot0028.png" "./gganim_plot0029.png" "./gganim_plot0030.png"
#>  [31] "./gganim_plot0031.png" "./gganim_plot0032.png" "./gganim_plot0033.png"
#>  [34] "./gganim_plot0034.png" "./gganim_plot0035.png" "./gganim_plot0036.png"
#>  [37] "./gganim_plot0037.png" "./gganim_plot0038.png" "./gganim_plot0039.png"
#>  [40] "./gganim_plot0040.png" "./gganim_plot0041.png" "./gganim_plot0042.png"
#>  [43] "./gganim_plot0043.png" "./gganim_plot0044.png" "./gganim_plot0045.png"
#>  [46] "./gganim_plot0046.png" "./gganim_plot0047.png" "./gganim_plot0048.png"
#>  [49] "./gganim_plot0049.png" "./gganim_plot0050.png" "./gganim_plot0051.png"
#>  [52] "./gganim_plot0052.png" "./gganim_plot0053.png" "./gganim_plot0054.png"
#>  [55] "./gganim_plot0055.png" "./gganim_plot0056.png" "./gganim_plot0057.png"
#>  [58] "./gganim_plot0058.png" "./gganim_plot0059.png" "./gganim_plot0060.png"
#>  [61] "./gganim_plot0061.png" "./gganim_plot0062.png" "./gganim_plot0063.png"
#>  [64] "./gganim_plot0064.png" "./gganim_plot0065.png" "./gganim_plot0066.png"
#>  [67] "./gganim_plot0067.png" "./gganim_plot0068.png" "./gganim_plot0069.png"
#>  [70] "./gganim_plot0070.png" "./gganim_plot0071.png" "./gganim_plot0072.png"
#>  [73] "./gganim_plot0073.png" "./gganim_plot0074.png" "./gganim_plot0075.png"
#>  [76] "./gganim_plot0076.png" "./gganim_plot0077.png" "./gganim_plot0078.png"
#>  [79] "./gganim_plot0079.png" "./gganim_plot0080.png" "./gganim_plot0081.png"
#>  [82] "./gganim_plot0082.png" "./gganim_plot0083.png" "./gganim_plot0084.png"
#>  [85] "./gganim_plot0085.png" "./gganim_plot0086.png" "./gganim_plot0087.png"
#>  [88] "./gganim_plot0088.png" "./gganim_plot0089.png" "./gganim_plot0090.png"
#>  [91] "./gganim_plot0091.png" "./gganim_plot0092.png" "./gganim_plot0093.png"
#>  [94] "./gganim_plot0094.png" "./gganim_plot0095.png" "./gganim_plot0096.png"
#>  [97] "./gganim_plot0097.png" "./gganim_plot0098.png" "./gganim_plot0099.png"
#> [100] "./gganim_plot0100.png"
#> attr(,"frame_vars")
#>     frame nframes progress frame_time
#> 1       1     100     0.01   18.23924
#> 2       2     100     0.02   18.54158
#> 3       3     100     0.03   18.84392
#> 4       4     100     0.04   19.14626
#> 5       5     100     0.05   19.44860
#> 6       6     100     0.06   19.75094
#> 7       7     100     0.07   20.05328
#> 8       8     100     0.08   20.35562
#> 9       9     100     0.09   20.65796
#> 10     10     100     0.10   20.96030
#> 11     11     100     0.11   21.26264
#> 12     12     100     0.12   21.56498
#> 13     13     100     0.13   21.86732
#> 14     14     100     0.14   22.16966
#> 15     15     100     0.15   22.47200
#> 16     16     100     0.16   22.77434
#> 17     17     100     0.17   23.07668
#> 18     18     100     0.18   23.37902
#> 19     19     100     0.19   23.68136
#> 20     20     100     0.20   23.98370
#> 21     21     100     0.21   24.28604
#> 22     22     100     0.22   24.58838
#> 23     23     100     0.23   24.89072
#> 24     24     100     0.24   25.19306
#> 25     25     100     0.25   25.49540
#> 26     26     100     0.26   25.79774
#> 27     27     100     0.27   26.10008
#> 28     28     100     0.28   26.40242
#> 29     29     100     0.29   26.70476
#> 30     30     100     0.30   27.00710
#> 31     31     100     0.31   27.30944
#> 32     32     100     0.32   27.61178
#> 33     33     100     0.33   27.91412
#> 34     34     100     0.34   28.21646
#> 35     35     100     0.35   28.51880
#> 36     36     100     0.36   28.82114
#> 37     37     100     0.37   29.12348
#> 38     38     100     0.38   29.42582
#> 39     39     100     0.39   29.72815
#> 40     40     100     0.40   30.03049
#> 41     41     100     0.41   30.33283
#> 42     42     100     0.42   30.63517
#> 43     43     100     0.43   30.93751
#> 44     44     100     0.44   31.23985
#> 45     45     100     0.45   31.54219
#> 46     46     100     0.46   31.84453
#> 47     47     100     0.47   32.14687
#> 48     48     100     0.48   32.44921
#> 49     49     100     0.49   32.75155
#> 50     50     100     0.50   33.05389
#> 51     51     100     0.51   33.35623
#> 52     52     100     0.52   33.65857
#> 53     53     100     0.53   33.96091
#> 54     54     100     0.54   34.26325
#> 55     55     100     0.55   34.56559
#> 56     56     100     0.56   34.86793
#> 57     57     100     0.57   35.17027
#> 58     58     100     0.58   35.47261
#> 59     59     100     0.59   35.77495
#> 60     60     100     0.60   36.07729
#> 61     61     100     0.61   36.37963
#> 62     62     100     0.62   36.68197
#> 63     63     100     0.63   36.98431
#> 64     64     100     0.64   37.28665
#> 65     65     100     0.65   37.58899
#> 66     66     100     0.66   37.89133
#> 67     67     100     0.67   38.19367
#> 68     68     100     0.68   38.49601
#> 69     69     100     0.69   38.79835
#> 70     70     100     0.70   39.10069
#> 71     71     100     0.71   39.40303
#> 72     72     100     0.72   39.70537
#> 73     73     100     0.73   40.00771
#> 74     74     100     0.74   40.31005
#> 75     75     100     0.75   40.61239
#> 76     76     100     0.76   40.91473
#> 77     77     100     0.77   41.21707
#> 78     78     100     0.78   41.51941
#> 79     79     100     0.79   41.82175
#> 80     80     100     0.80   42.12409
#> 81     81     100     0.81   42.42643
#> 82     82     100     0.82   42.72877
#> 83     83     100     0.83   43.03111
#> 84     84     100     0.84   43.33345
#> 85     85     100     0.85   43.63578
#> 86     86     100     0.86   43.93812
#> 87     87     100     0.87   44.24046
#> 88     88     100     0.88   44.54280
#> 89     89     100     0.89   44.84514
#> 90     90     100     0.90   45.14748
#> 91     91     100     0.91   45.44982
#> 92     92     100     0.92   45.75216
#> 93     93     100     0.93   46.05450
#> 94     94     100     0.94   46.35684
#> 95     95     100     0.95   46.65918
#> 96     96     100     0.96   46.96152
#> 97     97     100     0.97   47.26386
#> 98     98     100     0.98   47.56620
#> 99     99     100     0.99   47.86854
#> 100   100     100     1.00   48.17088
#>                                                                                frame_source
#> 1   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0001.png
#> 2   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0002.png
#> 3   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0003.png
#> 4   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0004.png
#> 5   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0005.png
#> 6   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0006.png
#> 7   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0007.png
#> 8   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0008.png
#> 9   C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0009.png
#> 10  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0010.png
#> 11  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0011.png
#> 12  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0012.png
#> 13  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0013.png
#> 14  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0014.png
#> 15  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0015.png
#> 16  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0016.png
#> 17  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0017.png
#> 18  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0018.png
#> 19  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0019.png
#> 20  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0020.png
#> 21  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0021.png
#> 22  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0022.png
#> 23  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0023.png
#> 24  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0024.png
#> 25  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0025.png
#> 26  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0026.png
#> 27  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0027.png
#> 28  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0028.png
#> 29  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0029.png
#> 30  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0030.png
#> 31  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0031.png
#> 32  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0032.png
#> 33  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0033.png
#> 34  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0034.png
#> 35  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0035.png
#> 36  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0036.png
#> 37  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0037.png
#> 38  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0038.png
#> 39  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0039.png
#> 40  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0040.png
#> 41  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0041.png
#> 42  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0042.png
#> 43  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0043.png
#> 44  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0044.png
#> 45  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0045.png
#> 46  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0046.png
#> 47  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0047.png
#> 48  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0048.png
#> 49  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0049.png
#> 50  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0050.png
#> 51  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0051.png
#> 52  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0052.png
#> 53  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0053.png
#> 54  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0054.png
#> 55  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0055.png
#> 56  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0056.png
#> 57  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0057.png
#> 58  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0058.png
#> 59  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0059.png
#> 60  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0060.png
#> 61  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0061.png
#> 62  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0062.png
#> 63  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0063.png
#> 64  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0064.png
#> 65  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0065.png
#> 66  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0066.png
#> 67  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0067.png
#> 68  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0068.png
#> 69  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0069.png
#> 70  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0070.png
#> 71  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0071.png
#> 72  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0072.png
#> 73  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0073.png
#> 74  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0074.png
#> 75  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0075.png
#> 76  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0076.png
#> 77  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0077.png
#> 78  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0078.png
#> 79  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0079.png
#> 80  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0080.png
#> 81  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0081.png
#> 82  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0082.png
#> 83  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0083.png
#> 84  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0084.png
#> 85  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0085.png
#> 86  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0086.png
#> 87  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0087.png
#> 88  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0088.png
#> 89  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0089.png
#> 90  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0090.png
#> 91  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0091.png
#> 92  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0092.png
#> 93  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0093.png
#> 94  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0094.png
#> 95  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0095.png
#> 96  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0096.png
#> 97  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0097.png
#> 98  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0098.png
#> 99  C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0099.png
#> 100 C:\\Users\\ROBIND~1\\AppData\\Local\\Temp\\RtmpeKJspk\\142874e07428/gganim_plot0100.png